ml/requirements.txt
joblib
keras
matplotlib
numpy
opencv_python
opencv_python_headless
pandas
Pillow
streamlit
streamlit_option_menu
tensorflow
uvicorn
fastapi
